refactor: use black and add annotation types

+++ src/staff/views.py
@@ -1,10 +1,12 @@
 import datetime
+from typing import Optional, Dict, List
 
 from django.contrib import messages
 from django.contrib.auth.decorators import permission_required
 from django.contrib.auth.models import Group
 from django.core.paginator import Paginator
-from django.db.models import OuterRef, Subquery, Count
+from django.db.models import OuterRef, Subquery, Count, QuerySet
+from django.http import HttpRequest, HttpResponse, QueryDict
 from django.shortcuts import render, redirect, get_object_or_404
 from django.utils import timezone
 from django.views.decorators.http import require_POST
@@ -44,8 +46,7 @@ from .services.permission_service import get_individually_assigned_permits
 
 
 @permission_required("staff.dashboard_view", raise_exception=False)
-def index(request):
-    title = "Главная страница"
+def index(request: HttpRequest) -> HttpResponse:
     today = datetime.date.today()
     completed_orders_today = number_of_completed_orders_today()
     orders_per_month = number_of_orders_per_month(today)
@@ -62,7 +63,7 @@ def index(request):
         request,
         "staff/dashboard.html",
         {
-            "title": title,
+            "title": "Главная страница",
             "number_of_completed_orders_today": completed_orders_today,
             "number_of_orders_per_month": orders_per_month,
             "summ_bill": summ,
@@ -75,7 +76,7 @@ def index(request):
 
 
 @permission_required("staff.manage_permission", raise_exception=True)
-def staff_group_list(request):
+def staff_group_list(request: HttpRequest) -> HttpResponse:
     groups = Group.objects.annotate(user_count=Count("user")).prefetch_related(
         "permissions"
     )
@@ -88,7 +89,7 @@ def staff_group_list(request):
 
 
 @permission_required("staff.manage_permission", raise_exception=True)
-def staff_group_edit(request, pk=None):
+def staff_group_edit(request: HttpRequest, pk: Optional[int] = None) -> HttpResponse:
     group = get_object_or_404(Group, pk=pk) if pk else None
     form = GroupForm(request.POST or None, instance=group)
     if request.method == "POST" and form.is_valid():
@@ -107,8 +108,8 @@ def staff_group_edit(request, pk=None):
 
 
 @permission_required("staff.manage_permission", raise_exception=True)
-def staff_list(request):
-    staffs = CustomUser.objects.filter(is_staff=True).prefetch_related(
+def staff_list(request: HttpRequest) -> HttpResponse:
+    staffs: QuerySet[CustomUser] = CustomUser.objects.filter(is_staff=True).prefetch_related(
         "groups", "user_permissions"
     )
     return render(
@@ -119,7 +120,7 @@ def staff_list(request):
 
 
 @permission_required("staff.manage_permission", raise_exception=True)
-def edit_staff_permissions(request, user_id):
+def edit_staff_permissions(request: HttpRequest, user_id: int) -> HttpResponse:
     user = get_object_or_404(CustomUser, id=user_id)
 
     if request.method == "POST":
@@ -141,7 +142,7 @@ def edit_staff_permissions(request, user_id):
 
     permissions = get_individually_assigned_permits()
 
-    permissions_by_app = {}
+    permissions_by_app: Dict = {}
     for perm in permissions:
         app_label = perm.content_type.app_label
         if app_label not in permissions_by_app:
@@ -163,8 +164,7 @@ def edit_staff_permissions(request, user_id):
 
 
 @permission_required("staff.manage_permission", raise_exception=True)
-def create_staff_user(request):
-    title = "Отправка приглашения сотруднику"
+def create_staff_user(request: HttpRequest) -> HttpResponse:
     if request.method == "POST":
         form = AdminCreateUserForm(request.POST)
         if form.is_valid():
@@ -172,13 +172,14 @@ def create_staff_user(request):
             return redirect("main_page")
     form = AdminCreateUserForm()
     return render(
-        request, "staff/create_staff_user.html", {"form": form, "title": title}
+        request,
+        "staff/create_staff_user.html",
+        {"form": form, "title": "Отправка приглашения сотруднику"},
     )
 
 
 @permission_required("cosmetics_shop.view_product", raise_exception=True)
-def products(request):
-    title = "Товары"
+def products(request: HttpRequest) -> HttpResponse:
     products = (
         Product.objects.all()
         .order_by("-id")
@@ -186,12 +187,14 @@ def products(request):
         .select_related("brand")
     )
 
-    query_params = request.GET.copy()
-    for key in list(query_params.keys()):
-        if not query_params[key].strip():
-            query_params.pop(key)
-    if request.GET.urlencode() != query_params.urlencode():
-        return redirect(f"{request.path}?{query_params.urlencode()}")
+    query_params: Optional[QueryDict] = request.GET.copy()
+    if query_params:
+        for key in list(query_params.keys()):
+            value = query_params.get(key, "")
+            if not value or not value.strip():
+                query_params.pop(key)
+        if request.GET.urlencode() != query_params.urlencode():
+            return redirect(f"{request.path}?{query_params.urlencode()}")
 
     form = ProductStuffFilterForm(request.GET or None)
     form_stock = FilterStockForm(request.GET or None)
@@ -224,14 +227,14 @@ def products(request):
 
     paginator = Paginator(products, 20)
     page_number = request.GET.get("page")
-    products = paginator.get_page(page_number)
+    page = paginator.get_page(page_number)
 
     return render(
         request,
         "staff/products.html",
         {
-            "title": title,
-            "products": products,
+            "title": "Товары",
+            "products": page,
             "form": form,
             "form_stock": form_stock,
         },
@@ -239,10 +242,10 @@ def products(request):
 
 
 @permission_required("cosmetics_shop.view_product", raise_exception=True)
-def product_card(request, product_id):
+def product_card(request: HttpRequest, product_id: int) -> HttpResponse:
     product = get_object_or_404(Product, pk=product_id)
     title = product.name
-    tags = product.tags.all()
+    tags: QuerySet[Tag] = product.tags.all()
     return render(
         request,
         "staff/product_card.html",
@@ -255,8 +258,7 @@ def product_card(request, product_id):
 
 
 @permission_required("cosmetics_shop.add_product", raise_exception=True)
-def create_products(request):
-    title = "Создание карточки товара"
+def create_products(request: HttpRequest) -> HttpResponse:
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -268,15 +270,14 @@ def create_products(request):
         request,
         "staff/create_product.html",
         {
-            "title": title,
+            "title": "Создание карточки товара",
             "form": form,
         },
     )
 
 
 @permission_required("cosmetics_shop.change_product", raise_exception=True)
-def edit_products(request, product_id):
-    title = "Изменение товара"
+def edit_products(request: HttpRequest, product_id: int) -> HttpResponse:
     product = get_object_or_404(Product, pk=product_id)
     if request.method == "POST":
         form = ProductForm(
@@ -290,7 +291,7 @@ def edit_products(request, product_id):
         request,
         "staff/edit_product.html",
         {
-            "title": title,
+            "title": "Изменение товара",
             "form": form,
             "product": product,
         },
@@ -299,17 +300,16 @@ def edit_products(request, product_id):
 
 @require_POST
 @permission_required("cosmetics_shop.delete_product", raise_exception=True)
-def delete_product(request):
+def delete_product(request: HttpRequest) -> HttpResponse:
     product_id = request.POST.get("product_id")
-    product = Product.objects.get(id=product_id)
+    product: Product = Product.objects.get(id=product_id)
     product.is_active = False
     product.save()
     return redirect("products")
 
 
 @permission_required("cosmetics_shop.view_order", raise_exception=True)
-def orders(request):
-    title = "Список заказов"
+def orders(request: HttpRequest) -> HttpResponse:
     latest_status_subquery = OrderStatusLog.objects.filter(
         order=OuterRef("order")
     ).order_by("-changed_at")
@@ -338,7 +338,7 @@ def orders(request):
                 request,
                 "staff/orders.html",
                 {
-                    "title": title,
+                    "title": "Список заказов",
                     "form": form,
                     "status": latest_statuses,
                 },
@@ -354,7 +354,7 @@ def orders(request):
         request,
         "staff/orders.html",
         {
-            "title": title,
+            "title": "Список заказов",
             "form": form,
             "status": latest_statuses,
         },
@@ -362,9 +362,9 @@ def orders(request):
 
 
 @permission_required("cosmetics_shop.view_order", raise_exception=True)
-def order_info(request, order_code):
+def order_info(request: HttpRequest, order_code: int) -> HttpResponse:
     title = f"Заказ {order_code}"
-    order = Order.objects.get(code=order_code)
+    order: Order = Order.objects.get(code=order_code)
     order_items = OrderItem.objects.filter(order=order)
     if request.method == "POST":
         form = OrderStatusForm(request.POST, instance=order, user=request.user)
@@ -408,8 +408,7 @@ def order_info(request, order_code):
 
 
 @permission_required("cosmetics_shop.view_brand", raise_exception=True)
-def brands_list(request):
-    title = "Список брендов"
+def brands_list(request: HttpRequest) -> HttpResponse:
     objects = Brand.objects.all()
 
     permissions = get_permissions(request, objects)
@@ -422,7 +421,7 @@ def brands_list(request):
         request,
         "staff/lists_page.html",
         {
-            "title": title,
+            "title": "Список брендов",
             "objects": objects,
             "permissions": permissions,
         },
@@ -430,73 +429,69 @@ def brands_list(request):
 
 
 @permission_required("cosmetics_shop.view_category", raise_exception=True)
-def categories_list(request):
-    title = "Список категорий"
-    objects = Category.objects.all()
+def categories_list(request: HttpRequest) -> HttpResponse:
+    objects: QuerySet[Category] = Category.objects.all()
 
     permissions = get_permissions(request, objects)
 
     paginator = Paginator(objects, 20)
     page_number = request.GET.get("page")
-    objects = paginator.get_page(page_number)
+    page = paginator.get_page(page_number)
 
     return render(
         request,
         "staff/lists_page.html",
         {
-            "objects": objects,
-            "title": title,
+            "objects": page,
+            "title": "Список категорий",
             "permissions": permissions,
         },
     )
 
 
 @permission_required("cosmetics_shop.view_tag", raise_exception=True)
-def tags_list(request):
-    title = "Список тегов"
-    objects = Tag.objects.all()
+def tags_list(request: HttpRequest) -> HttpResponse:
+    objects: QuerySet[Tag] = Tag.objects.all()
 
     permissions = get_permissions(request, objects)
 
     paginator = Paginator(objects, 20)
     page_number = request.GET.get("page")
-    objects = paginator.get_page(page_number)
+    page = paginator.get_page(page_number)
 
     return render(
         request,
         "staff/lists_page.html",
         {
-            "objects": objects,
-            "title": title,
+            "title": "Список тегов",
+            "objects": page,
             "permissions": permissions,
         },
     )
 
 
 @permission_required("cosmetics_shop.view_groupproduct", raise_exception=True)
-def groups_list(request):
-    title = " Список групп"
-    objects = GroupProduct.objects.all()
+def groups_list(request: HttpRequest) -> HttpResponse:
+    objects: QuerySet[GroupProduct] = GroupProduct.objects.all()
     permissions = get_permissions(request, objects)
 
     paginator = Paginator(objects, 20)
     page_number = request.GET.get("page")
-    objects = paginator.get_page(page_number)
+    page = paginator.get_page(page_number)
 
     return render(
         request,
         "staff/lists_page.html",
         {
-            "objects": objects,
-            "title": title,
+            "title": "Список групп",
+            "objects": page,
             "permissions": permissions,
         },
     )
 
 
 @permission_required("cosmetics_shop.add_category", raise_exception=True)
-def create_categories(request):
-    title = "Создание категории"
+def create_categories(request: HttpRequest) -> HttpResponse:
     if request.method == "POST":
         form = CategoryForm(request.POST)
         if form.is_valid():
@@ -508,15 +503,14 @@ def create_categories(request):
         request,
         "staff/create_page.html",
         {
-            "title": title,
+            "title": "Создание категории",
             "form": form,
         },
     )
 
 
 @permission_required("cosmetics_shop.add_groupproduct", raise_exception=True)
-def create_groups(request):
-    title = "Создание группы"
+def create_groups(request: HttpRequest) -> HttpResponse:
     if request.method == "POST":
         form = GroupProductForm(request.POST)
         if form.is_valid():
@@ -528,15 +522,14 @@ def create_groups(request):
         request,
         "staff/create_page.html",
         {
-            "title": title,
+            "title": "Создание группы",
             "form": form,
         },
     )
 
 
 @permission_required("cosmetics_shop.add_brand", raise_exception=True)
-def create_brands(request):
-    title = "Создание бренда"
+def create_brands(request: HttpRequest) -> HttpResponse:
     if request.method == "POST":
         form = BrandForm(request.POST)
         if form.is_valid():
@@ -548,15 +541,14 @@ def create_brands(request):
         request,
         "staff/create_page.html",
         {
-            "title": title,
+            "title": "Создание бренда",
             "form": form,
         },
     )
 
 
 @permission_required("cosmetics_shop.add_tag", raise_exception=True)
-def create_tags(request):
-    title = "Создание тега"
+def create_tags(request: HttpRequest) -> HttpResponse:
     if request.method == "POST":
         form = TagForm(request.POST)
         if form.is_valid():
@@ -568,7 +560,7 @@ def create_tags(request):
         request,
         "staff/create_page.html",
         {
-            "title": title,
+            "title": "Создание тега",
             "form": form,
         },
     )
@@ -576,42 +568,54 @@ def create_tags(request):
 
 @require_POST
 @permission_required("cosmetics_shop.delete_tag", raise_exception=True)
-def delete_tags(request):
+def delete_tags(request: HttpRequest) -> HttpResponse:
     tag_id = request.POST.get("id")
-    tag = get_object_or_404(Tag, id=tag_id)
-    tag.delete()
+    if tag_id is not None:
+        tag = get_object_or_404(Tag, id=tag_id)
+        tag.delete()
+    else:
+        messages.error(request, "Не удалось удалить тэг")
     return redirect("tags_list")
 
 
 @require_POST
 @permission_required("cosmetics_shop.delete_category", raise_exception=True)
-def delete_categories(request):
+def delete_categories(request: HttpRequest) -> HttpResponse:
     category_id = request.POST.get("id")
-    category = get_object_or_404(Category, id=category_id)
-    category.delete()
+    if category_id is not None:
+        category = get_object_or_404(Category, id=category_id)
+        category.delete()
+    else:
+        messages.error(request, "Не удалось удалить категорию")
     return redirect("categories_list")
 
 
 @require_POST
 @permission_required("cosmetics_shop.delete_groupproduct", raise_exception=True)
-def delete_groups(request):
+def delete_groups(request: HttpRequest) -> HttpResponse:
     group_id = request.POST.get("id")
-    group = get_object_or_404(GroupProduct, id=group_id)
-    group.delete()
+    if group_id is not None:
+        group = get_object_or_404(GroupProduct, id=group_id)
+        group.delete()
+    else:
+        messages.error(request, "Не удалось удалить группу")
     return redirect("groups_list")
 
 
 @require_POST
 @permission_required("cosmetics_shop.delete_brand", raise_exception=True)
-def delete_brands(request):
+def delete_brands(request: HttpRequest) -> HttpResponse:
     brand_id = request.POST.get("id")
-    brand = get_object_or_404(Brand, id=brand_id)
-    brand.delete()
+    if brand_id is not None:
+        brand = get_object_or_404(Brand, id=brand_id)
+        brand.delete()
+    else:
+        messages.error(request, "Не удалось удалить бренд")
     return redirect("brands_list")
 
 
 @permission_required("cosmetics_shop.change_category", raise_exception=True)
-def edit_categories(request, pk):
+def edit_categories(request: HttpRequest, pk: int) -> HttpResponse:
     category = get_object_or_404(Category, id=pk)
     title = f"Изменение категории: {category.name}"
 
@@ -633,7 +637,7 @@ def edit_categories(request, pk):
 
 
 @permission_required("cosmetics_shop.change_groupproduct", raise_exception=True)
-def edit_groups(request, pk):
+def edit_groups(request: HttpRequest, pk: int) -> HttpResponse:
     group = get_object_or_404(GroupProduct, id=pk)
     title = f"Изменение группы: {group.name}"
 
@@ -655,7 +659,7 @@ def edit_groups(request, pk):
 
 
 @permission_required("cosmetics_shop.change_brand", raise_exception=True)
-def edit_brands(request, pk):
+def edit_brands(request: HttpRequest, pk: int) -> HttpResponse:
     brand = get_object_or_404(Brand, id=pk)
     title = f"Изменение бренда: {brand.name}"
     if request.method == "POST":
@@ -676,7 +680,7 @@ def edit_brands(request, pk):
 
 
 @permission_required("cosmetics_shop.change_tag", raise_exception=True)
-def edit_tags(request, pk):
+def edit_tags(request: HttpRequest, pk: int) -> HttpResponse:
     tag = get_object_or_404(Tag, id=pk)
     title = f"Изменение тега: {tag.name}"
